fix(auth): move logout into settings
