+++ apps/web/src/components/workspace/left-panel.tsx
@@ -17,7 +17,6 @@ import {
   Plugs,
   Plus,
   Robot,
-  SignOut,
   SlidersHorizontal,
   Sun,
   Warning,
@@ -155,7 +154,6 @@ type LeftPanelProps = {
   onNavigateSettings: () => void;
   onNavigateConnectors?: () => void;
   onNavigateProviders?: () => void;
-  onLogout?: () => void;
 
   // Sessions
   sessions: WorkspaceSession[];
@@ -274,7 +272,6 @@ function MinifiedLeftPanel({
   onExpandWithSection,
   onSyncComplete,
   onNavigateSettings,
-  onLogout,
 }: {
   slug: string;
   configChangePending?: boolean;
@@ -289,7 +286,6 @@ function MinifiedLeftPanel({
   onExpandWithSection: (section: "chats" | "knowledge" | "experts" | "skills") => void;
   onSyncComplete?: (status: SyncKbResult["status"]) => void;
   onNavigateSettings: () => void;
-  onLogout?: () => void;
 }) {
   const {
     themes,
@@ -533,22 +529,6 @@ function MinifiedLeftPanel({
           </TooltipTrigger>
           <TooltipContent side="right">Settings</TooltipContent>
         </Tooltip>
-
-        {onLogout ? (
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <button
-                type="button"
-                onClick={onLogout}
-                className="flex h-8 w-8 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:bg-foreground/5 hover:text-foreground"
-                aria-label="Log out"
-              >
-                <SignOut size={16} weight="bold" />
-              </button>
-            </TooltipTrigger>
-            <TooltipContent side="right">Log out</TooltipContent>
-          </Tooltip>
-        ) : null}
       </div>
     </TooltipProvider>
   );
@@ -573,7 +553,6 @@ export function LeftPanel({
   onNavigateSettings,
   onNavigateConnectors,
   onNavigateProviders,
-  onLogout,
   sessions,
   activeSessionId,
   hasMoreSessions,
@@ -625,7 +604,6 @@ export function LeftPanel({
          onExpandWithSection={handleExpandWithSection}
         onSyncComplete={onSyncComplete}
         onNavigateSettings={onNavigateSettings}
-        onLogout={onLogout}
       />
     );
   }
@@ -649,7 +627,6 @@ export function LeftPanel({
       onNavigateSettings={onNavigateSettings}
       onNavigateConnectors={onNavigateConnectors}
       onNavigateProviders={onNavigateProviders}
-      onLogout={onLogout}
       sessions={sessions}
       activeSessionId={activeSessionId}
       hasMoreSessions={hasMoreSessions}
@@ -695,7 +672,6 @@ function ExpandedLeftPanel({
   onNavigateSettings,
   onNavigateConnectors,
   onNavigateProviders,
-  onLogout,
   onRestartConfig,
   sessions,
   activeSessionId,
@@ -1595,22 +1571,6 @@ function ExpandedLeftPanel({
             </TooltipTrigger>
             <TooltipContent side="top">Settings</TooltipContent>
           </Tooltip>
-
-          {onLogout ? (
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <button
-                  type="button"
-                  onClick={onLogout}
-                  className="flex h-7 w-7 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:bg-foreground/5 hover:text-foreground"
-                  aria-label="Log out"
-                >
-                  <SignOut size={14} weight="bold" />
-                </button>
-              </TooltipTrigger>
-              <TooltipContent side="top">Log out</TooltipContent>
-            </Tooltip>
-          ) : null}
         </TooltipProvider>
       </div>
